neatogen genXConstraints: take 'n' as a 'size_t'
Pays down some warnings from prior commits.

+++ lib/vpsc/csolve_VPSC.cpp
@@ -36,10 +36,10 @@ VPSC* newIncVPSC(int n, Variable* vs[], int m, Constraint* cs[]) {
 	return new IncVPSC(n,vs,m,cs);
 }
 
-int genXConstraints(int n, boxf *bb, Variable **vs, Constraint ***cs,
+int genXConstraints(size_t n, boxf *bb, Variable **vs, Constraint ***cs,
                     bool transitiveClosure) {
 	std::vector<Rectangle> rs;
-	for(int i=0;i<n;i++) {
+	for (size_t  i = 0; i < n; i++) {
 		rs.emplace_back(bb[i].LL.x,bb[i].UR.x,bb[i].LL.y,bb[i].UR.y);
 	}
 	const int m = generateXConstraints(rs, vs, *cs, transitiveClosure);
